server/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron_descriptor==2.0.6
decorator==5.2.1
Django==5.0.3
django-celery-beat==2.8.1
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-timezone-field==7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
executing==2.2.1
gunicorn==22.0.0
ipdb==0.13.13
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.3.6
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg==3.2.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.8.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
